web/requirements.txt
Django>=5.2.3
gunicorn
psycopg2-binary
cryptography
python-dotenv
dj-database-url
djangorestframework
drf-spectacular 
cryptography          # requis par PyJWT pour RS256
PyJWT>=2.8            # ajout 👉  indispensable